11. August 2021, Digital Breakthrough, Text & Image Classification/requirements.txt
albumentations==0.5.2
tensorflow==2.5.0
numpy>=1.19.5
torch>=1.7.1
pandas==1.1.5
torchvision>=0.8.2
opencv-python
PyYAML
tqdm==4.56.0
scikit-image
scikit-learn
scipy
matplotlib
python-json-logger>=0.1.11
jupyterlab
seaborn
#grad-cam
#ttach
transformers
omegaconf